Include/methodobject.h: Move Py_LIMITED_API declarations to cpython/
